refactor(workflow/status): collapse runtime service and WorkflowStatePort
Absorbs the pass-through methods from PrevcStatusRuntimeService into free
functions in statusTransitions.ts, eliminating the statusManager -> runtime
service -> transitions triple indirection. statusManager now calls pure
helpers directly.

Collapses the WorkflowStatePort fake DI (single-implementation interface)
into a direct dependency on HarnessWorkflowStateService. workflowStatePort.ts
is retained as a deprecated type alias to keep workflow/orchestrator.ts
compiling without touching it in this stream.

Also drops the per-instance cachedStatus: reads are a single small JSON file,
and a process-local cache loses value once MCP + CLI can both write. An
mtime-based cache can be reintroduced if profiling justifies it.

Unifies the sync and async read paths through a single finalizeRead helper.
