+++ src/services/harness/workflowStateService.ts
@@ -9,7 +9,6 @@
 import * as fs from 'fs-extra';
 import * as path from 'path';
 import type { PrevcStatus } from '../../workflow/types';
-import type { WorkflowStatePort } from '../../workflow/status/workflowStatePort';
 
 export interface HarnessWorkflowStateServiceOptions {
   contextPath: string;
@@ -31,7 +30,7 @@ export interface HarnessWorkflowRecord {
   binding: WorkflowHarnessBinding | null;
 }
 
-export class HarnessWorkflowStateService implements WorkflowStatePort {
+export class HarnessWorkflowStateService {
   constructor(private readonly options: HarnessWorkflowStateServiceOptions) {}
 
   private get contextPath(): string {
