+++ src/workflow/status/statusManager.ts
@@ -1,8 +1,15 @@
 /**
  * PREVC Status Manager
  *
- * Coordinates canonical PREVC workflow state using a persistence port,
- * a runtime transition service, and a legacy YAML migrator.
+ * Coordinates canonical PREVC workflow state using a persistence service,
+ * pure transition helpers, and a legacy YAML migrator.
+ *
+ * No in-memory cache is maintained: canonical state lives in a single small
+ * JSON file under `.context/harness/workflows/`, reads are cheap, and an
+ * mtime-based cache would add invalidation complexity that loses value once
+ * multiple processes (MCP + CLI) can write concurrently. If profiling ever
+ * shows read contention, reintroduce caching keyed on the canonical file's
+ * mtime — not on a stale in-process value.
  */
 
 import type {
@@ -21,19 +28,42 @@ import type {
 import { PREVC_PHASE_ORDER, getPhaseDefinition } from '../phases';
 import { getRoleConfig } from '../prevcConfig';
 import { createInitialStatus } from './templates';
-import type { WorkflowStatePort } from './workflowStatePort';
-import { PrevcStatusRuntimeService } from './runtimeStatusService';
+import type { HarnessWorkflowStateService } from '../../services/harness/workflowStateService';
+import {
+  addHistoryEntry as addHistoryEntryOnStatus,
+  addStepHistoryEntry as addStepHistoryEntryOnStatus,
+  applyWorkflowSettings,
+  approvePlanOnStatus,
+  completeStatusPhaseTransition,
+  getActiveRole as getActiveRoleFromStatus,
+  getCurrentPhase as getCurrentPhaseFromStatus,
+  getNextPhaseForStatus,
+  isWorkflowComplete,
+  markPlanCreatedOnStatus,
+  markStatusPhaseComplete,
+  normalizeLoadedStatus,
+  type StepHistoryInput,
+  transitionStatusToPhase,
+  updateAgent as updateAgentOnStatus,
+  updatePhase as updatePhaseOnStatus,
+  updateRole as updateRoleOnStatus,
+} from './statusTransitions';
 import { LegacyWorkflowStatusMigrator } from '../legacy/legacyWorkflowStatusMigrator';
 import { getDefaultSettings } from '../gates';
 
+interface StatusReadResult {
+  status: PrevcStatus;
+  fromLegacy: boolean;
+  changed: boolean;
+  hasLegacyProjection: boolean;
+}
+
 export class PrevcStatusManager {
-  private cachedStatus: PrevcStatus | null = null;
-  private readonly statusRuntime = new PrevcStatusRuntimeService();
   private readonly legacyMigrator: LegacyWorkflowStatusMigrator;
 
   constructor(
     contextPath: string,
-    private readonly workflowState: WorkflowStatePort
+    private readonly workflowState: HarnessWorkflowStateService
   ) {
     this.legacyMigrator = new LegacyWorkflowStatusMigrator(contextPath);
   }
@@ -42,51 +72,48 @@ export class PrevcStatusManager {
     return (await this.workflowState.exists()) || (await this.legacyMigrator.exists());
   }
 
-  private async readStatus(): Promise<{ status: PrevcStatus; fromLegacy: boolean; changed: boolean; hasLegacyProjection: boolean }> {
-    const hasCanonicalState = await this.workflowState.exists();
-    const hasLegacyProjection = await this.legacyMigrator.exists();
-
-    if (!hasCanonicalState && !hasLegacyProjection) {
+  private finalizeRead(
+    raw: PrevcStatus | null,
+    fromLegacy: boolean,
+    hasLegacyProjection: boolean
+  ): StatusReadResult {
+    if (!raw) {
       throw new Error('Workflow status not found. Run "workflow init" first.');
     }
 
-    const status = hasCanonicalState
-      ? await this.workflowState.load()
-      : await this.legacyMigrator.load();
-
-    const normalized = this.statusRuntime.normalizeLoadedStatus(status);
+    const normalized = normalizeLoadedStatus(raw);
     return {
       status: normalized.status,
-      fromLegacy: !hasCanonicalState,
+      fromLegacy,
       changed: normalized.changed,
       hasLegacyProjection,
     };
   }
 
-  private readStatusSync(): { status: PrevcStatus; fromLegacy: boolean; changed: boolean; hasLegacyProjection: boolean } {
-    const canonicalExists = this.workflowState.existsSync();
-    const legacyExists = this.legacyMigrator.existsSync();
-    let status: PrevcStatus | null = null;
-    let fromLegacy = false;
+  private async readStatus(): Promise<StatusReadResult> {
+    const hasCanonicalState = await this.workflowState.exists();
+    const hasLegacyProjection = await this.legacyMigrator.exists();
 
-    if (canonicalExists) {
-      status = this.workflowState.loadSync();
-    } else if (legacyExists) {
-      status = this.legacyMigrator.loadSync();
-      fromLegacy = true;
-    }
+    const raw = hasCanonicalState
+      ? await this.workflowState.load()
+      : hasLegacyProjection
+        ? await this.legacyMigrator.load()
+        : null;
 
-    if (!status) {
-      throw new Error('Workflow status not found. Run "workflow init" first.');
-    }
+    return this.finalizeRead(raw, !hasCanonicalState && hasLegacyProjection, hasLegacyProjection);
+  }
 
-    const normalized = this.statusRuntime.normalizeLoadedStatus(status);
-    return {
-      status: normalized.status,
-      fromLegacy,
-      changed: normalized.changed,
-      hasLegacyProjection: legacyExists,
-    };
+  private readStatusSync(): StatusReadResult {
+    const hasCanonicalState = this.workflowState.existsSync();
+    const hasLegacyProjection = this.legacyMigrator.existsSync();
+
+    const raw = hasCanonicalState
+      ? this.workflowState.loadSync()
+      : hasLegacyProjection
+        ? this.legacyMigrator.loadSync()
+        : null;
+
+    return this.finalizeRead(raw, !hasCanonicalState && hasLegacyProjection, hasLegacyProjection);
   }
 
   private async persistStatus(status: PrevcStatus): Promise<void> {
@@ -95,45 +122,30 @@ export class PrevcStatusManager {
   }
 
   async load(): Promise<PrevcStatus> {
-    if (this.cachedStatus) {
-      return this.cachedStatus;
-    }
-
     const read = await this.readStatus();
     if (read.changed || read.fromLegacy || read.hasLegacyProjection) {
       await this.persistStatus(read.status);
     }
-
-    this.cachedStatus = read.status;
-    return this.cachedStatus;
+    return read.status;
   }
 
   loadSync(): PrevcStatus {
-    if (this.cachedStatus) {
-      return this.cachedStatus;
-    }
-
-    const read = this.readStatusSync();
-    this.cachedStatus = read.status;
-    return this.cachedStatus;
+    return this.readStatusSync().status;
   }
 
   async save(status: PrevcStatus): Promise<void> {
-    const normalized = this.statusRuntime.normalizeLoadedStatus(status);
+    const normalized = normalizeLoadedStatus(status);
     await this.persistStatus(normalized.status);
-    this.cachedStatus = normalized.status;
   }
 
   async remove(): Promise<void> {
     await this.workflowState.remove();
     await this.legacyMigrator.remove();
-    this.cachedStatus = null;
   }
 
   async archive(name: string): Promise<void> {
     await this.workflowState.archive(name);
     await this.legacyMigrator.archive(name);
-    this.cachedStatus = null;
   }
 
   async create(options: {
@@ -155,64 +167,64 @@ export class PrevcStatusManager {
 
   async updatePhase(phase: PrevcPhase, update: PhaseUpdate): Promise<void> {
     const status = await this.load();
-    this.statusRuntime.updatePhase(status, phase, update);
+    updatePhaseOnStatus(status, phase, update);
     await this.save(status);
   }
 
   async updateRole(role: PrevcRole, update: RoleUpdate): Promise<void> {
     const status = await this.load();
-    this.statusRuntime.updateRole(status, role, update);
+    updateRoleOnStatus(status, role, update);
     await this.save(status);
   }
 
   async updateAgent(agentName: string, update: AgentUpdate): Promise<void> {
     const status = await this.load();
-    this.statusRuntime.updateAgent(status, agentName, update);
+    updateAgentOnStatus(status, agentName, update);
     await this.save(status);
   }
 
   async transitionToPhase(phase: PrevcPhase): Promise<void> {
     const status = await this.load();
-    this.statusRuntime.transitionToPhase(status, phase);
+    transitionStatusToPhase(status, phase);
     await this.save(status);
   }
 
   async markPhaseComplete(phase: PrevcPhase, outputs?: string[]): Promise<void> {
     const status = await this.load();
-    this.statusRuntime.markPhaseComplete(status, phase, outputs);
+    markStatusPhaseComplete(status, phase, outputs);
     await this.save(status);
   }
 
   async completePhaseTransition(outputs?: string[]): Promise<PrevcPhase | null> {
     const status = await this.load();
-    const nextPhase = this.statusRuntime.completePhaseTransition(status, outputs);
+    const nextPhase = completeStatusPhaseTransition(status, outputs);
     await this.save(status);
     return nextPhase;
   }
 
   async getCurrentPhase(): Promise<PrevcPhase> {
     const status = await this.load();
-    return this.statusRuntime.getCurrentPhase(status);
+    return getCurrentPhaseFromStatus(status);
   }
 
   async getActiveRole(): Promise<PrevcRole | null> {
     const status = await this.load();
-    return this.statusRuntime.getActiveRole(status);
+    return getActiveRoleFromStatus(status);
   }
 
   async getNextPhase(): Promise<PrevcPhase | null> {
     const status = await this.load();
-    return this.statusRuntime.getNextPhase(status);
+    return getNextPhaseForStatus(status);
   }
 
   async isComplete(): Promise<boolean> {
     const status = await this.load();
-    return this.statusRuntime.isComplete(status);
+    return isWorkflowComplete(status);
   }
 
   async setSettings(settings: Partial<WorkflowSettings>): Promise<WorkflowSettings> {
     const status = await this.load();
-    const nextSettings = this.statusRuntime.setSettings(status, settings);
+    const nextSettings = applyWorkflowSettings(status, settings);
     await this.save(status);
     return nextSettings;
   }
@@ -224,13 +236,13 @@ export class PrevcStatusManager {
 
   async markPlanCreated(planSlug: string): Promise<void> {
     const status = await this.load();
-    this.statusRuntime.markPlanCreated(status, planSlug);
+    markPlanCreatedOnStatus(status, planSlug);
     await this.save(status);
   }
 
   async approvePlan(approver: PrevcRole | string, notes?: string): Promise<PlanApproval> {
     const status = await this.load();
-    const approval = this.statusRuntime.approvePlan(status, approver, notes);
+    const approval = approvePlanOnStatus(status, approver, notes);
     await this.save(status);
     return approval;
   }
@@ -242,21 +254,13 @@ export class PrevcStatusManager {
 
   async addHistoryEntry(entry: Omit<ExecutionHistoryEntry, 'timestamp'>): Promise<void> {
     const status = await this.load();
-    this.statusRuntime.addHistoryEntry(status, entry);
+    addHistoryEntryOnStatus(status, entry);
     await this.save(status);
   }
 
-  async addStepHistoryEntry(entry: {
-    action: 'step_started' | 'step_completed' | 'step_skipped';
-    plan: string;
-    planPhase: string;
-    stepIndex: number;
-    stepDescription?: string;
-    output?: string;
-    notes?: string;
-  }): Promise<void> {
+  async addStepHistoryEntry(entry: StepHistoryInput): Promise<void> {
     const status = await this.load();
-    this.statusRuntime.addStepHistoryEntry(status, entry);
+    addStepHistoryEntryOnStatus(status, entry);
     await this.save(status);
   }
 
@@ -267,7 +271,7 @@ export class PrevcStatusManager {
 
   async getSummary(): Promise<WorkflowSummary> {
     const status = await this.load();
-    const isComplete = this.statusRuntime.isComplete(status);
+    const complete = isWorkflowComplete(status);
 
     let completedPhases = 0;
     let totalPhases = 0;
@@ -290,7 +294,7 @@ export class PrevcStatusManager {
         total: totalPhases,
         percentage: Math.round((completedPhases / totalPhases) * 100),
       },
-      isComplete,
+      isComplete: complete,
       startedAt: status.project.started,
     };
   }
